Accounting complete

+++ src/main/java/ait/cohort46/post/dto/CommentDto.java
@@ -1,20 +1,19 @@
 package ait.cohort46.post.dto;
 
+import java.time.LocalDateTime;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class CommentDto {
-    private String user;
+	private String user;
     private String message;
     private LocalDateTime dateCreated;
-    private Long likes;
-
+    private Integer likes;
 }
